content/smil/test/test_smilGetStartTime.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>Test for getStartTime Behavior </title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
<svg id="svg" xmlns="http://www.w3.org/2000/svg" width="120px" height="120px">
  <circle cx="20" cy="20" r="15" fill="blue">
    <animate attributeName="cx" attributeType="XML" 
      from="20" to="100" begin="indefinite" dur="1s" id="anim"/>
  </circle>
</svg>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[
/** Test for getStartTime Behavior  **/

/* Global Variables */
var svg = document.getElementById("svg");
var anim = document.getElementById("anim");

SimpleTest.waitForExplicitFinish();

function main() {
  // indefinite 
  is(anim.getStartTime(), 0, "Unexpected start time with indefinite begin");

  // 1s
  anim.setAttribute("begin", "1s");
  is(anim.getStartTime(), 1, "Unexpected start time with begin=1s");

  // We have to be careful here when choosing a negative time that we choose
  // a time that will create an interval that reaches past t=0 as SMIL has
  // special rules for throwing away intervals that end before t=0
  anim.setAttribute("begin", "-0.5s");
  is(anim.getStartTime(), -0.5, "Unexpected start time with begin=-0.5s");

  // Now test over the lifetime of the animation when there are multiple
  // intervals
  anim.setAttribute("begin", "1s; 3s");
  is(anim.getStartTime(), 1, "Unexpected start time before first interval");
  
  svg.setCurrentTime(1);
  is(anim.getStartTime(), 1,
     "Unexpected start time at start of first interval");

  svg.setCurrentTime(1.5);
  is(anim.getStartTime(), 1, "Unexpected start time during first interval");

  svg.setCurrentTime(2);
  is(anim.getStartTime(), 1, "Unexpected start time after first interval");

  svg.setCurrentTime(3);
  is(anim.getStartTime(), 3, "Unexpected start time during second interval");

  svg.setCurrentTime(4);
  is(anim.getStartTime(), 3, "Unexpected start time after second interval");

  SimpleTest.finish();
}

window.addEventListener("load", main, false);
]]>
</script>
</pre>
</body>
</html>
